ApplicationLibrary/mpu6050/requirements.txt
paho-mqtt==1.6.1
smbus
